Implemented following basics
Buggy and hacky. The user object that is stored in the session is not the same kind of user object that is stored
in the user database, which caused some problems.
Bug: 'Followers' and 'following' lists do not get updated correctly--user B has user A as both follower
and following, but user B doesn't show up on either of user A's lists
Unfinished: not allowing you to follow a person more than once